+++ src/main/kotlin/racingcar/Car.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 interface Car {
