+++ src/main/kotlin/racingcar/RacingCarGame.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 class RacingCarGame {
