+++ src/main/kotlin/racingcar/RacingCarGameInputView.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 class RacingCarGameInputView {
