+++ src/main/kotlin/racingcar/RacingCarGameResultView.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 class RacingCarGameResultView {
