+++ src/main/kotlin/racingcar/RacingGameCarFactory.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 object RacingGameCarFactory {
