+++ src/test/kotlin/racingcar/RacingGameCarFactoryTest.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
